Potential fix for code scanning alert no. 51: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using the size_t type, which has a larger range than int.

 - Cast one of the operands (height or width) to size_t before performing the multiplication.
 - Update the assignment to data_size to use the casted multiplication result.

+++ libwebp-1.0.3/src/utils/quant_levels_utils.c
@@ -30,7 +30,7 @@ int QuantizeLevels(uint8_t* const data, int width, int height,
   int q_level[NUM_SYMBOLS] = { 0 };
   double inv_q_level[NUM_SYMBOLS] = { 0 };
   int min_s = 255, max_s = 0;
-  const size_t data_size = height * width;
+  const size_t data_size = (size_t)height * width;
   int i, num_levels_in, iter;
   double last_err = 1.e38, err = 0.;
   const double err_threshold = ERROR_THRESHOLD * data_size;
